+++ examples/template-hydrogen-default/src/App.server.jsx
@@ -12,7 +12,7 @@ export default function App({log, pages, ...serverState}) {
   return (
     <Suspense fallback={<LoadingFallback />}>
       <ShopifyProvider shopifyConfig={shopifyConfig}>
-        <AppClient helmetContext={serverState.helmetContext}>
+        <AppClient>
           <DefaultSeo />
           <DefaultRoutes
             pages={pages}