+++ packages/hydrogen/src/client.ts
@@ -5,5 +5,5 @@ export * from './foundation/useShop';
 export * from './foundation/ServerStateProvider';
 export {ShopifyProviderClient as ShopifyProvider} from './foundation/ShopifyProvider/ShopifyProvider.client';
 export {ServerStateRouter} from './foundation/Router/ServerStateRouter.client';
-export {Helmet, HelmetProvider} from 'react-helmet-async';
+export {Helmet} from './foundation/Helmet';
 export * from './utilities';