+++ packages/hydrogen/src/components/Seo/DescriptionSeo.client.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Helmet} from 'react-helmet-async';
+import {Helmet} from '../../client';
 
 import type {Description} from './types';