+++ packages/hydrogen/src/entry-server.tsx
@@ -14,11 +14,10 @@ import {
 import {getErrorMarkup} from './utilities/error';
 import {defer} from './utilities/defer';
 import type {ImportGlobEagerOutput, ServerHandler} from './types';
-import type {FilledContext} from 'react-helmet-async';
 import {Html} from './framework/Hydration/Html';
 import {Renderer, Hydrator, Streamer} from './types';
 import {ServerComponentResponse} from './framework/Hydration/ServerComponentResponse.server';
-import {ServerComponentRequest} from './framework/Hydration/ServerComponentRequest.server';
+import type {ServerComponentRequest} from './framework/Hydration/ServerComponentRequest.server';
 import {getCacheControlHeader} from './framework/cache';
 import {ServerRequestProvider} from './foundation/ServerRequestProvider';
 import type {ServerResponse} from 'http';
@@ -30,6 +29,7 @@ import {ServerStateProvider} from './foundation/ServerStateProvider';
 import {renderToReadableStream as rscRenderToReadableStream} from '@shopify/hydrogen/vendor/react-server-dom-vite/writer.browser.server';
 // @ts-ignore
 import {createFromReadableStream} from '@shopify/hydrogen/vendor/react-server-dom-vite';
+import type {RealHelmetData} from './foundation/Helmet/Helmet';
 
 declare global {
   // This is provided by a Vite plugin
@@ -60,7 +60,7 @@ const renderHydrogen: ServerHandler = (App, {pages}) => {
     const log = getLoggerFromContext(request);
     const state = {pathname: url.pathname, search: url.search};
 
-    const {ReactApp, helmetContext, componentResponse} = buildReactApp({
+    const {ReactApp, componentResponse} = buildReactApp({
       App,
       state,
       request,
@@ -99,9 +99,10 @@ const renderHydrogen: ServerHandler = (App, {pages}) => {
     }
 
     headers['Content-type'] = HTML_CONTENT_TYPE;
-    const params = {url, ...extractHeadElements(helmetContext)};
+    const {bodyAttributes, htmlAttributes, ...head} = extractHeadElements(
+      request.ctx.helmet
+    );
 
-    const {bodyAttributes, htmlAttributes, ...head} = params;
     head.script = (head.script || '') + flightContainer({init: true, nonce});
 
     html = html
@@ -160,7 +161,9 @@ const renderHydrogen: ServerHandler = (App, {pages}) => {
     const ReactAppSSR = (
       <Html template={template} htmlAttrs={{lang: 'en'}}>
         <ServerRequestProvider request={request} isRSC={false}>
-          <RscConsumer />
+          <ServerStateProvider serverState={state} setServerState={() => {}}>
+            <RscConsumer />
+          </ServerStateProvider>
         </ServerRequestProvider>
       </Html>
     );
@@ -512,12 +515,10 @@ function buildReactApp({
   isRSC?: boolean;
   pages?: ImportGlobEagerOutput;
 }) {
-  const helmetContext = {} as FilledContext;
   const componentResponse = new ServerComponentResponse();
   const hydrogenServerProps = {
     request,
     response: componentResponse,
-    helmetContext,
     log,
   };
 
@@ -535,12 +536,10 @@ function buildReactApp({
     return <React.Suspense fallback={null}>{AppContent}</React.Suspense>;
   };
 
-  return {helmetContext, ReactApp, componentResponse};
+  return {ReactApp, componentResponse};
 }
 
-function extractHeadElements(helmetContext: FilledContext) {
-  const {helmet} = helmetContext;
-
+function extractHeadElements({context: {helmet}}: RealHelmetData) {
   return helmet
     ? {
         base: helmet.base.toString(),